combine unit tests into one file
